+++ db/migrate/20161223120627_add_is_verified_to_client.rb
@@ -1,0 +1,5 @@
+class AddIsVerifiedToClient < ActiveRecord::Migration[5.0]
+  def change
+    add_column :clients, :verified, :boolean, default: false
+  end
+end